pyproject.toml
[tool.poetry]
name = "dm-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.7"
pre-commit = "^3.1.1"
python-dotenv = "^1.0.0"
model-bakery = "^1.10.1"
pytest-dotenv = "^0.5.2"
pytest-cov = "^4.0.0"
djangorestframework = "^3.14.0"
latexvm = "^0.1.6"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
psycopg2-binary = "^2.9.5"
pytest-django = "^4.5.2"

[tool.poetry.group.group.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
